Crate ntools

A modular toolkit of fast and reliable libraries for neutronics analysis

Tony’s neutronics toolbox

A modular toolkit of fast and reliable libraries for neutronics analysis

The full library documentation is published here

This is a pre-release version for testing and development, so breaking changes will often occur until a stable 1.0 release.

Library overview

The ntools toolkit contains a collection of mostly modular libraries for common fusion neutronics tasks and analysis.

Crate	Description

fispact	Analysis tools for FISPACT-II inventory calculations

iaea	Module for interacting with the IAEA decay data API

mesh	MCNP mesh tally operations and file parsing

mctal	MCNP MCTAL file parsing and tally data

posvol	Se/deserialiser for UKAEA CuV posvol binaries

utils	Common utilities and extension traits

weights	Tools for MCNP weight window operations

wwgen	Weight window generation methods for MCNP

Command line tools built on these core libraries are maintained in their own repositories.

Modular crates

The ntools crates are included as dependencies through feature flags. Specify "full" to include everything.

[dependencies]
# via SSH
ntools = { git = "ssh://git@github.com/repositony/ntools.git", features = ["full"] }

# via HTTPS
ntools = { git = "https://github.com/repositony/ntools.git", features = ["full"] }

It is recommended that users are more selective to avoid compiling unnecessary dependencies.

For example, if only the fispact and iaea crates are needed:

[dependencies]
# via SSH
ntools = { git = "ssh://git@github.com/repositony/ntools.git", features = ["fispact", "iaea"] }

# via HTTPS
ntools = { git = "https://github.com/repositony/ntools.git", features = ["fispact", "iaea"] }

Documentation and Tests

To reproduce the full library documentation seen here:

cargo doc --workspace --no-deps --features full

To run all tests for all modules, use the --workspace flag.

cargo test --workspace

Modules

fispact

Analysis tools for FISPACT-II inventory calculations

iaea

Query decay data directly from the IAEA chart of nuclides

mctal

Module for reading MCNP MCTAL files

mesh

Mesh tally tools and file parsing

posvol

Module for working with UKAEA CuV posvol binaries

weights

Tools for MCNP weight window operations

wwgen

Weight window generation methods for MCNP