ntools::iaea

Function load_all

pub fn load_all(rad_type: RadType) -> &'static HashMap<String, Vec<Record>>

Load all nuclides from pre-fetched data

This will load all nuclides that were pre-processed into binary files from the full IAEA chart of nuclides.

Specify the type of decay data with the appropriate RadType to get the full dataset.

// Load all records for all pre-fetched nuclides
let decay_data: &HashMap<String, RecordSet> = load_all(RadType::Gamma);

The returned hashmap is a dictionary of key value pairs where:

key : Nuclide name, e.g. “Co60”

value : List of all matching Records